[FIX] NRE & Code Readability

+++ OpenNos.GameObject/ArenaMember.cs
@@ -10,8 +10,8 @@ namespace OpenNos.GameObject
     public class ArenaMember
     {
         public ClientSession Session { get; set; }
-        public double? GroupId { get; set; }
+        public long? GroupId { get; set; }
         public EventType ArenaType { get; set; }
-        public int Time { get;  set; }
+        public int Time { get; set; }
     }
 }
